+++ app/src/main/java/org/hwyl/sexytopo/control/util/StationRotator.java
@@ -6,7 +6,7 @@ import org.hwyl.sexytopo.model.survey.Station;
 
 public class StationRotator {
 
-    public static Station rotate(Station station, double angle) {
+    public static Station rotate(Station station, float angle) {
         Station rotated = new Station(station.getName());
 
         // this would be a lot cleaner with the Java 8 streaming API :'(
