+++ app/src/main/java/org/hwyl/sexytopo/model/sketch/DeletedDetail.java
@@ -29,7 +29,7 @@ public class DeletedDetail extends SketchDetail {
     }
 
     @Override
-    public double getDistanceFrom(Coord2D point) {
+    public float getDistanceFrom(Coord2D point) {
         return deletedDetail.getDistanceFrom(point);
     }
