+++ app/src/main/java/org/hwyl/sexytopo/model/sketch/SinglePositionDetail.java
@@ -19,7 +19,7 @@ public abstract class SinglePositionDetail extends SketchDetail {
     }
 
     @Override
-    public double getDistanceFrom(Coord2D point) {
+    public float getDistanceFrom(Coord2D point) {
         return Space2DUtils.getDistance(point, position);
     }
 }
